Splitting the python client testsuite.
Into a library and the binary. In some environments, creating the
connection cannot be done easily.

Also adding some test cases for exceptions.
